Ignore the weight of worn objects by default
The Spectrum PAW documentation claims that the weight of worn objects counts
towards your carrying capacity, but testing shows it actually isn't.

I haven't been able to create a container in the Spetrum version of The PAW, so
wasn't able to check the behaviour for items carried within a worn container
(like an anvil contained in a backpack).  I've just assumed they still count
towards weight.

I've enabled this by default (for compatibility with The PAW), but made it
configurable from the variable "worn_items_have_weight" (for compatibility with
existing ngPAWS games).